move all the formatters to their own subdirectory
