+++ src/smc-project/formatters/html-format.ts
@@ -2,7 +2,6 @@ const { writeFile, readFile } = require("fs");
 const tmp = require("tmp");
 const { callback } = require("awaiting");
 const { spawn } = require("child_process");
-// const { replace_all } = require("smc-util/misc");
 
 interface ParserOptions {
   parser: string;
